✨(front) add tabs in the DashboardVideoLiveControlPane
We add tabs in the instructor's dashboard to enable more screen views.
A default configuration tab is added. A new tab appears to show the
attendances of the live.
